+++ config/params.config.js
@@ -2,4 +2,4 @@ module.exports = {
   adminEmail: 'admin@example.com',
   senderEmail: 'noreply@example.com',
   senderName: 'Example.com mailer'
-}
+};