+++ config/web.config.js
@@ -1,6 +1,8 @@
-const log4js = require('log4js')
-const db = require('./db.config')
-const params = require('./params.config')
+const log4js = require('log4js');
+const db = require('./db.config');
+const params = require('./params.config');
+const Config = require('../core/lib/Config');
+
 const webOptions = {
   appName: 'Gracile',
   key: 'Z3JhY2lsZQ',
@@ -12,12 +14,17 @@ const webOptions = {
     template: 'twig'
   },
   params: params
-}
+};
 if (process.env.NODE_ENV === 'development') {
-  webOptions.components.bootstrap.push('debug')
+  webOptions.components.bootstrap.push('debug');
 }
-const logger = log4js.getLogger(webOptions.appName)
-logger.level = 'debug'
-webOptions.components.logger = logger
+const logger = log4js.getLogger(webOptions.appName);
+logger.level = 'debug';
+webOptions.components.logger = logger;
 
-module.exports = webOptions
+const WebConfig = new Config({
+  private: ['key'],
+  config: webOptions
+});
+
+module.exports = WebConfig;