+++ controllers/docs.controller.js
@@ -1,19 +1,19 @@
-const Controller = require('../core/lib/Controller')
+const Controller = require('../core/lib/Controller');
 
 class DocsController extends Controller {
   main (req, res, next) {
-    res.send('New controller was created')
+    res.send('New controller was created');
 
     if (next) {
       // Middleware
-      next() // Continue stack
+      next(); // Continue stack
     } else {
       // Controller
-      res.end() // End stack
+      res.end(); // End stack
     }
   }
 }
 
-const docsController = new DocsController()
+const docsController = new DocsController();
 
-module.exports = docsController
+module.exports = docsController;