+++ prototypes/router/router.prototype.js
@@ -3,12 +3,12 @@
  * @desc {name} router of server
  * */
 // Declaring router
-const Router = require('express').Router();
+const Router = require('express').Router()
 
 // Router methods
-Router.all('/', (req, res)=>{
-  res.send('New router');
-});
+Router.all('/', (req, res) => {
+  res.send('New router')
+})
 
 // Export
-module.exports = Router;
+module.exports = Router