+++ utils/debug.js
@@ -1,7 +1,7 @@
-const webConfig = require('../config/web.config')
-const { logger } = webConfig.components
+const webConfig = require('../config/web.config');
+const { logger } = webConfig.components;
 
 module.exports = (req, res, next) => {
-  logger.debug('Request on: ' + req.url)
-  next()
-}
+  logger.debug('Request on: ' + req.url);
+  next();
+};